# public/api/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Evaly API - Système de Gestion des Étudiants",
        "description": "API REST pour la gestion des étudiants, authentification et profils dans le système Evaly",
        "contact": {
            "email": "support@evaly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3002",
            "description": "Serveur API Evaly"
        }
    ],
    "paths": {
...
        "/api/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Récupérer la liste des étudiants",
                "description": "Récupère la liste paginée des étudiants avec filtres optionnels",
                "operationId": "5a54d6df45ca455ad022d2f7b9da9731",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Terme de recherche (nom, prénom, email, numéro étudiant)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "class_id",
                        "in": "query",
                        "description": "ID de la classe pour filtrer",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Statut actif (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des étudiants récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Créer un nouvel étudiant",
                "description": "Crée un nouvel étudiant avec les données fournies",
                "operationId": "aa0044bcc7e0cbaa736f0c20a4a6765c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "student_number": {
                                        "type": "string",
                                        "example": "STU202400001"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+33123456789"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Rue de la Paix, Paris"
                                    },
                                    "parent_contact": {
                                        "type": "string",
                                        "example": "parent@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "motdepasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Étudiant créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Récupérer un étudiant par ID",
                "description": "Récupère les détails d'un étudiant spécifique",
                "operationId": "5e2013c1980b929b399fa6e9296534b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'étudiant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'étudiant récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Étudiant non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Mettre à jour un étudiant",
                "description": "Met à jour les informations d'un étudiant existant",
                "operationId": "35f77190e85ebd1a750bc9a367dfe651",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'étudiant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+33123456789"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Rue de la Paix, Paris"
                                    },
                                    "parent_contact": {
                                        "type": "string",
                                        "example": "parent@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Étudiant mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/students/import/{classId}": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Importer des étudiants depuis un fichier Excel",
                "description": "Importe une liste d'étudiants depuis un fichier Excel et les assigne à une classe",
                "operationId": "1f810408942cbd1fc77e3ec057301874",
                "parameters": [
                    {
                        "name": "classId",
                        "in": "path",
                        "description": "ID de la classe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Fichier Excel contenant les données des étudiants",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import réalisé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/students/{id}/assign-class/{classId}": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Assigner un étudiant à une classe",
                "description": "Assigne un étudiant existant à une classe",
                "operationId": "6cec3fbcb963f493006f45f144c222a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'étudiant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "classId",
                        "in": "path",
                        "description": "ID de la classe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Étudiant assigné à la classe avec succès"
                    }
                }
            }
        },
        "/api/classes/{classId}/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Récupérer les étudiants d'une classe",
                "description": "Récupère tous les étudiants assignés à une classe spécifique",
                "operationId": "60f3bef6a90d62057cc01c92063d2e2a",
                "parameters": [
                    {
                        "name": "classId",
                        "in": "path",
                        "description": "ID de la classe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des étudiants de la classe récupérée avec succès"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez le token Bearer obtenu lors de la connexion",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}